requirements.txt
numpy
pandas
torch
langchain
pypdf
openai 
langchain-openai
langchain_pinecone
langchain[docarray]
langchain_community
sentence-transformers
docarray
pydantic==1.10.8
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
python-dotenv
-e .